requirements.txt
alembic==1.6.5
autopep8==1.5.7
Babel==2.9.0
backports.entry-points-selectable==1.1.0
click==8.0.1
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-Moment==1.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
isort==5.9.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
platformdirs==2.0.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.20
toml==0.10.2
virtualenv==20.5.0
Werkzeug==2.0.1
WTForms==2.3.3
